complete ft_printf

+++ Milestone.1/ft_printf/ft_putnbr_fd_base.c
@@ -6,18 +6,18 @@
 /*   By: skamijo <skamijo@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/01 10:42:01 by skamijo           #+#    #+#             */
-/*   Updated: 2024/11/10 03:19:28 by skamijo          ###   ########.fr       */
+/*   Updated: 2025/02/13 01:20:05 by skamijo          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libftprintf.h"
+#include "ft_printf.h"
 
-int	ft_putnbr_fd_base(long long n, char *base, int base_num, int fd)
+int	ft_putnbr_fd_base(unsigned int n, char *base, int base_num, int fd)
 {
 	int	ret;
 
 	ret = 0;
-	if (n >= base_num)
+	if (n >= (unsigned int)base_num)
 	{
 		ret += ft_putnbr_fd_base(n / base_num, base, base_num, fd);
 		ret += ft_putnbr_fd_base(n % base_num, base, base_num, fd);
